1.0.5

+++ src/stdlib/timestamp/mod.rs
@@ -6,8 +6,8 @@ use rhai::plugin::*;
 #[export_module]
 pub mod timestamp_module {
 
-    pub fn timestamp_ms() -> i64 {
-    	SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_millis() as i64
+    pub fn timestamp_ms() -> f64 {
+    	SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_millis() as f64
     }
 }
